extract the struct Ws2812ColorOrder into its own file
